Update dependencies, update project setup, include Run.ps1 script
