+++ README.md
@@ -1,14 +1,10 @@
-# Seq.Input.AzureEventHub
+# Seq.Input.Azure.EventHub
 
-[![Build Status](https://dev.azure.com/hinnipipe/Github%20Pipeline/_apis/build/status/Hinni.Seq.Input.AzureEventHub?branchName=master)](https://dev.azure.com/hinnipipe/Github%20Pipeline/_build/latest?definitionId=4&branchName=master)
 [![GitHub release](https://img.shields.io/github/release/Hinni/Seq.Input.AzureEventHub.svg)](https://github.com/Hinni/Seq.Input.AzureEventHub/releases)
 [![NuGet](https://img.shields.io/nuget/v/Seq.Input.AzureEventHub.svg)](https://www.nuget.org/packages/Seq.Input.AzureEventHub/)
 
-A Seq custom input that pulls messages from Azure EventHub. **Requires Seq 5.1+ and currently works only on docker and not on Windows instances**
-
-## Getting started
-
-The app is published to NuGet as [_Seq.Input.AzureEventHub_](https://nuget.org/packages/Seq.Input.AzureEventHub). Follow the instructions for [installing a Seq App](https://docs.getseq.net/docs/installing-seq-apps) and start an instance of the app, providing your Azure Event Hub details.
+A Seq custom input that pulls messages from Azure EventHub. Built by @Hinni, currently being
+experimentally updated to test with new dependencies and host versions.
 
 ## Sending events to the input
