+++ Seq.Input.AzureEventHub.sln
@@ -6,18 +6,18 @@ MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{A8975350-F515-4ABC-A6CB-9F732E7D391D}"
 	ProjectSection(SolutionItems) = preProject
 		.gitignore = .gitignore
-		azure-pipelines.yml = azure-pipelines.yml
 		LICENSE = LICENSE
 		README.md = README.md
+		Run.ps1 = Run.ps1
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Seq.Input.AzureEventHub", "src\Seq.Input.AzureEventHub\Seq.Input.AzureEventHub.csproj", "{D84C061E-1014-44E3-A1BA-21048F292709}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Seq.Input.Azure.EventHub", "src\Seq.Input.Azure.EventHub\Seq.Input.Azure.EventHub.csproj", "{D84C061E-1014-44E3-A1BA-21048F292709}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{5C9AEE60-0494-4CDD-8F47-874A69813F2D}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{64F3E611-7453-4EE2-93B1-1AC31F3C9958}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Seq.Input.AzureEventHub.TestConsole", "src\Seq.Input.AzureEventHub.TestConsole\Seq.Input.AzureEventHub.TestConsole.csproj", "{7FECE51A-F94C-490A-81AB-109B6245F19A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Seq.Input.Azure.EventHub.TestConsole", "src\Seq.Input.Azure.EventHub.TestConsole\Seq.Input.Azure.EventHub.TestConsole.csproj", "{7FECE51A-F94C-490A-81AB-109B6245F19A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
