+++ src/Seq.Input.Azure.EventHub.TestConsole/appsettings.json
@@ -1,0 +1,6 @@
+{
+  "AppSettings": {
+    "EventHubConnectionString": "Endpoint=sb://",
+    "EventHubName": "HubName",
+  }
+}
