--- src/Seq.Input.AzureEventHub.TestConsole/appsettings.json
@@ -1,9 +1,0 @@
-{
-  "AppSettings": {
-    "EventHubConnectionString": "Endpoint=sb://",
-    "EventHubName": "HubName",
-    "ConsumerGroupName": "$Default",
-    "StorageConnectionString": "DefaultEndpointsProtocol=https;AccountName=",
-    "StorageContainerName": "BlobName"
-  }
-}
